group9/smart-cart-system/fog-backend/requirements.txt
django
djangorestframework
psycopg2-binary
pandas
django-cors-headers
dj-database-url
djangorestframework-simplejwt
networkx
scipy
